refactor java project: 2015
