+++ 2015/java/aoc-2015/settings.gradle
@@ -1,0 +1,2 @@
+rootProject.name = 'aoc-2015'
+
