+++ 2015/java/aoc-2015/src/main/resources/day1/input.txt
@@ -1,0 +1,1 @@
+/Users/pete/code/aoc/2015/inputs/day1.txt
