--- 2015/java/day1/input.txt
@@ -1,1 +1,0 @@
-../../python/day1-not-quite-lisp/input.txt
