--- 2015/java/day2/input.txt
@@ -1,1 +1,0 @@
-../../python/day2/input.txt
