--- 2015/java/day3/input.txt
@@ -1,1 +1,0 @@
-../../python/day3/input.txt
